refactor: remove uint8clamped array
also moved all render code away from Canvas
